unordered sups: cleaner oportunistic swap approach
unordered sups and dups will now be identical to normal sups and dups,
rather than dups storing only one argument. to solve the quadratic
issue, we just perform oportunistic swaps: whenever we enter the aux
port of an unordered dup node that is connected to another unordered
dup, we swap both nodes. this will give "higher priority" to nodes we
accessed recently. this ensures that, in cases where there is an
accumulation of redundant dup nodes, these will be harmless, since we'll
push the relevant dup nodes closer to the duplicated values, and the
redundant ones will not be accessed.

NOTE: the code is currently completely duplicated for the implementation
of dups/sups and unordered sups/dups. I'll clean up later