HVM lazy evaluation propagates to Haskell side
Now, instead of expecting a normalized term, the extract/collapse
functions will call reduce layer by layer, with unsafeInterleaveIO. As
a result, laziness will propagate from Haskell to HVM, in the sense
that we will only compute interactions that are observable on Haskell
side. So, for example, if part of the term is discarded, the count
will also decrease, as the irrelevant interactions won't be mde.

Also, the Collapser will now use a priority queue, so that the list
of λ-Terms collapsed is visited from lowest to highest labels. This
has proven to be relevant for proof search. Other strategies may be
considered later, though.

I've also added the missing Collapse.hs file.