2-port dup nodes; reuse memory; enforce affinity
dup nodes are now stored in two ports. before reduction, we use the
first port to represent the value, and the second is just SUB. after
reduction, both ports become the substitutions.

we'll now reuse memory of some nodes, instead of allocating.

these changes increase performance by ~50% in some cases

I've also noticed that, sadly, we can't actually use variables
non-linearly on branches, as in:

λb ~x{
  #T: ... b ...
  #F: ... b ...
}

that's because, in certain superposed situations, we can actually reach
both branches, causing the 'b' pointer to appear in two places in
memory, which should never happen. as such, we need to linearize, as we
did before:

λb (~x{
  #T: λb ... b ...
  #F: λb ... b ...
} b)

because of that, HVM3 will now throw an error if a variable is used
non-linearly, even if across different branches.

I'll update the syntax to make this limitation more palatable

also, note that, now, the fast-compiled mode will sometimes perform more
interactions than the interpreted mode. that's because some
optimizations (like applying the λ's of a mat interaction) lose some
laziness (perhaps these λ's wouldn't be applied in a very superposed
situation). that's fine and I'll keep it like that. we could add a
"compiled but without optimizations" option later if we want a
consistent interaction count.