bugfix; 10 billion itr/s single-core!
we can not call 'reduce()' inside the hot-path attempt, otherwise this
can cause reduce to be called twice on the same term (in cases of undo),
which, in turn, can cause duplicated terms in memory.

by fixing this, the count.hvml test magically jumps to 10 billion MIPS
(about the same performance as Haskell), and the sum_range to 2139 MIPS
(almost 6x faster than Haskell!). of course, for the hot paths to kick,
we now need to place strict annotations correctly

I've also added more memore reuses and the nots.hvml bench is up to
91 MIPS, and the pseudo_metavars_factors.hvml is up to 100 MIPS, which
makes it 22x faster than the Haskell equivalent!

I've also attempted to introduce a freelist, but it didn't result in
significant speedups and I don't think it was worth the extra code
and operations.

I know I was supposed to rest but I had to do this btw, the new gpt-4o
model is really good, it translated a compiled hvm.c file (20K tokens)
to JS in a almost single pass, and that allowed me to find this issue