dyn sup/dup syntax; auto sup/dup syntax; faster collapser; allow cloning unboxed var; rename exampels
now '&' can be part of a variable name. variables that start with '&'
are assumed to be unboxed (i.e., numbers, nullary constructors, era)
and, as such, can be used multiple times.

now we can use '&a{x y}' and '&a{x y}=v body' and to create a dyn
sup/dup with '&a' label, as a syntax sugar for '@SUP(&a x y)' and
'@DUP(&a v λxλy(body))'

now, when we omit the label of a sup/dup node, HVML will generate
a static fresh name for you. ex: '&{1 2}' becomes '&8388608{1 2}'

the collapser has been fixed to use IntMap instead of lists, so that
large labels don't make it slow anymore

all the examples on book/ have been renamed and cleaned up