default cases on match (desugars to if-let)
HVM now has a new mode for Mat, named if-let:

~ x {
  #Foo{..}: main-case
  other: default-case
}

it behaves as an if-let: if the first constructor matches, return the
main case applied to the fields, otherwise, return default-case applied
to the original value.

when there is more than one case and a default, as in:

~ x {
  #Foo{..}: main-case
  #Bar{..}: main-case
  ...
  other: default-case
}

it gets desugared to a chain of if-lets.

if-let is more lightweight than the usual mat because it avoid storing
all cases in memory, but, when there are many cases covered, it might
be slower as it requires additional interactions. it also allows for
more concise code in some cases

if-let is not optimized at all by the compiler for now