fast DUP-REF and REF-SUP interactions (copy mode)
now function declarations can be tagged with a '!', which will make them
eligible for the copy interactions:

    @foo(&L{ax ay} b c ...)
    ----------------------- REF-SUP-COPY (when @L not in @foo)
    ! &L{bx by} = b
    ! &L{cx cy} = b
    ...
    &L{@foo(ax bx cx ...) @foo(ay by cy ...)}

    ! &L{x y} = @foo(a b c ...)
    --------------------------- DUP-REF-COPY (when &L not in @foo)
    ! &L{a0 a1} = a
    ! &L{b0 b1} = b
    ! &L{c0 c1} = c
    ...
    x <- @foo(a0 b0 c0 ...)
    y <- @foo(a1 b1 c1 ...)

these rules will allow REF nodes to be cloned and collapsed before
dereferencing their bodies, which can be useful in some cases. it can be
used to ensure the hot path of compiled functions is hit, even when the
argument is a superposition (because it would turn '@foo(&{#A #B})' into
'&0{@foo(#A) @foo(#B)}'. it is also useful when passing higher order
fns.  for example `&0{f0 f1} = @foo` would copy just a reference to
'@foo', instead of expanding and cloning its body, so that, when it is
actually called, it might also activate the hot paths. that said, this
interaction breaks optimality, so, for example, a function won't get
"auto memoized" when copy mode is enabled for it, and interaction count
can grow in many similar situations; in some cases, exponentially so.
overall this is good to give the programmer a control mechanism to
ensure hot paths on the C compiler are hit, but it should be used with
caution as to not lose optimality when it is desirable.