Dups with only one memory slot
After reducing a duplication, we'd leave two traces of the substitution (one for each variable)
and return one of those two duplicated values.

By the logic of reduce, the value that was left for the varaible that we just entered from
was never actually read since we always use the return. That means that we only need to write
for the other variable.

Before reducing, the second slot just had a placeholder value, meaning that we actually never need
2 slots for dups, only one.
```c
// Before
  Loc du0     = alloc_node(2);
  set(du0 + 0, bod);
  set(du0 + 1, term_new(SUB, 0, 0));
  ...
  sub(dup_loc + 0, term_new(LAM, 0, lm0));
  sub(dup_loc + 1, term_new(LAM, 0, lm1));
  sub(lam_loc + 0, term_new(SUP, dup_lab, su0));
  return term_rem_bit(got(dup_loc + dup_num));

// After
  Loc du0     = alloc_node(1);
  set(du0 + 0, bod);
  sub(lam_loc + 0, term_new(SUP, dup_lab, su0));
  if (term_tag(dup) == DP0) {
    sub(dup_loc + 0, term_new(LAM, 0, lm1));
    return term_new(LAM, 0, lm0);
  } else {
    sub(dup_loc + 0, term_new(LAM, 0, lm0));
    return term_new(LAM, 0, lm1);
  }
```

These changes lead to ~15% increase in MIPS and ~20% reduction in space for SupGen, obviously
it depends on the program. I'm also seeing a small improvement even in programs with very little
duplications (5-10%), I'm guessing the smaller space allows better cache or better code optimizations.